Increase performance of 2D TEF and mask empty bins
By introducing a simple if-condition in the double for-loop of the
sort_2dim function, the computation can be made much faster.  In the
2D-TEF diagram (e.g., with temperature and salinity), there are often
many empty bins with no data for a given T-S-combination.  By skipping
these bins, we save a lot of computations and thus time.  In test runs,
the computation time decreased by more than 50 %.  Furthermore, bins
that are empty at all time steps are now marked by NaN (instead of 0),
so that the 2D-TEF diagram shows the difference between bins that are
always empty and those with zero transport.